// website/types/athlete-prs.ts
export type AthletePRs = {
  athlete_name: string
  snatch_1rm: number | null
  snatch_2rm: number | null
  snatch_3rm: number | null
  clean_1rm: number | null
  clean_2rm: number | null
  clean_3rm: number | null
  jerk_1rm: number | null
  jerk_2rm: number | null
  jerk_3rm: number | null
  clean_jerk_1rm: number | null
  clean_jerk_2rm: number | null
  clean_jerk_3rm: number | null
  back_squat_1rm: number | null
  back_squat_2rm: number | null
  back_squat_3rm: number | null
  back_squat_4rm: number | null
  back_squat_5rm: number | null
  back_squat_6rm: number | null
  back_squat_7rm: number | null
  back_squat_8rm: number | null
  back_squat_9rm: number | null
  back_squat_10rm: number | null
  front_squat_1rm: number | null
  front_squat_2rm: number | null
  front_squat_3rm: number | null
  front_squat_4rm: number | null
  front_squat_5rm: number | null
  front_squat_6rm: number | null
  front_squat_7rm: number | null
  front_squat_8rm: number | null
  front_squat_9rm: number | null
  front_squat_10rm: number | null
  created_at?: string
  updated_at?: string
}